Rename abstract TachyonFileSystem to AbstractTachyonFileSystem and StreamingTachyonFileSystem to TachyonFileSystem.
