+++ clients/unshaded/src/main/java/tachyon/client/TachyonFS.java
@@ -39,7 +39,6 @@ import tachyon.TachyonURI;
 import tachyon.annotation.PublicApi;
 import tachyon.client.block.BlockStoreContext;
 import tachyon.client.file.FileSystemContext;
-import tachyon.client.file.StreamingTachyonFileSystem;
 import tachyon.client.table.RawTable;
 import tachyon.conf.TachyonConf;
 import tachyon.thrift.DependencyDoesNotExistException;
@@ -63,7 +62,7 @@ import tachyon.worker.WorkerClient;
  * Under the hood, this class maintains a MasterClientBase to talk to the master server and
  * WorkerClients to interact with different Tachyon workers.
  *
- * As of 0.8, replaced by {@link StreamingTachyonFileSystem}
+ * As of 0.8, replaced by {@link tachyon.client.file.TachyonFileSystem}
  */
 @PublicApi
 @Deprecated
