+++ clients/unshaded/src/main/java/tachyon/client/file/TachyonFileSystem.java
@@ -16,211 +16,152 @@
 package tachyon.client.file;
 
 import java.io.IOException;
-import java.util.List;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.google.common.base.Preconditions;
 
-import tachyon.Constants;
 import tachyon.TachyonURI;
-import tachyon.annotation.PublicApi;
+import tachyon.client.ClientOptions;
 import tachyon.client.FileSystemMasterClient;
 import tachyon.thrift.BlockInfoException;
+import tachyon.thrift.DependencyDoesNotExistException;
 import tachyon.thrift.FileAlreadyExistException;
 import tachyon.thrift.FileDoesNotExistException;
 import tachyon.thrift.FileInfo;
 import tachyon.thrift.InvalidPathException;
 
 /**
- * Tachyon File System client. This class should be used to interface with the Tachyon File
- * System master and supports all non IO operations. Implementing classes should provide their
- * own IO methods separate from the {@link TachyonFileSystemCore} interface.
+ * A TachyonFileSystem implementation including convenience methods as well as a streaming API to
+ * read and write files. Creating files is not supported by this TachyonFileSystem because the
+ * files should only be written once, thus getOutStream is sufficient for creating and writing to
+ * a file.
  */
-@PublicApi
-public abstract class TachyonFileSystem implements TachyonFileSystemCore {
-  private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
+public class TachyonFileSystem extends AbstractTachyonFileSystem {
 
-  /** The file system context which contains shared resources, such as the fs master client */
-  // TODO (calvin): Make this private after StreamingTachyonFileSystem is cleaned up
-  protected FileSystemContext mContext;
+  private static TachyonFileSystem sTachyonFileSystem;
 
-  /**
-   * Constructor, currently TachyonFileSystem does not retain any state
-   */
-  protected TachyonFileSystem() {
-    mContext = FileSystemContext.INSTANCE;
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public long create(TachyonURI path, long blockSize, boolean recursive) throws
-      BlockInfoException, FileAlreadyExistException, InvalidPathException, IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      return masterClient.createFile(path.getPath(), blockSize, recursive);
-    } finally {
-      mContext.releaseMasterClient(masterClient);
+  public static synchronized TachyonFileSystem get() {
+    if (sTachyonFileSystem == null) {
+      sTachyonFileSystem = new TachyonFileSystem();
     }
+    return sTachyonFileSystem;
+  }
+
+  private TachyonFileSystem() {
+    super();
   }
 
   /**
-   * {@inheritDoc}
-   *
-   * The delete will abort on a failure, but previous deletes (if deleting more than
-   * one file) that occurred will still be effective. The delete will only synchronously be
-   * propagated to the master. The file metadata will not be available after this call, but the data
-   * in Tachyon or under storage space may still reside until the delete is propagated and all
-   * current readers have relinquished their locks.
+   * Create is not supported in StreamingTachyonFileSystem. Files should only be written once.
+   * Use getOutStream instead to create files.
    */
   @Override
-  public void delete(TachyonFile file, boolean recursive) throws FileDoesNotExistException,
+  public long create(TachyonURI path, long blockSize, boolean recursive) {
+    throw new UnsupportedOperationException("Create is not supported, use getOutStream instead.");
+  }
+
+  /**
+   * Convenience function for delete recursively. This is the same as calling delete(file, true).
+   *
+   * @param file the TachyonFile to delete recursively
+   * @throws FileDoesNotExistException if the file does not exist in Tachyon space
+   * @throws IOException if the master cannot delete the file
+   */
+  public void delete(TachyonFile file) throws FileDoesNotExistException, IOException {
+    delete(file, true);
+  }
+
+  /**
+   * Convenience function for free recursively. This is the same as calling free(file, true).
+   *
+   * @param file the TachyonFile to free recursively
+   * @throws FileDoesNotExistException if the file does not exist in Tachyon space
+   * @throws IOException if the master cannot delete the file
+   */
+  public void free(TachyonFile file) throws FileDoesNotExistException,
       IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      masterClient.deleteFile(file.getFileId(), recursive);
-      LOG.info(
-          "Deleted file " + file.getFileId() + " from both Tachyon Storage and under file system");
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
+    free(file, true);
   }
 
   /**
-   * {@inheritDoc}
+   * Gets a {@link FileInStream} for the specified file. The stream's settings can be customized by
+   * setting the options parameter. The caller should close the stream after finishing the
+   * operations on it.
    *
-   * This method is asynchronous and will be propagated to the workers through their heartbeats.
+   * @param file the handler for the file.
+   * @param options the set of options specific to this operation.
+   * @return an input stream to read the file
+   * @throws FileDoesNotExistException if the file does not exist
+   * @throws IOException if the stream cannot be opened for some other reason
    */
-  @Override
-  public void free(TachyonFile file, boolean recursive) throws FileDoesNotExistException,
+  public FileInStream getInStream(TachyonFile file, ClientOptions options) throws IOException,
+      FileDoesNotExistException {
+    FileInfo info = getInfo(file);
+    Preconditions.checkState(!info.isIsFolder(), "Cannot read from a folder");
+    return new FileInStream(info, options);
+  }
+
+  /**
+   * Creates a file and gets the {@link FileOutStream} for the specified file. If the parent
+   * directories do not exist, they will be created. This should only be called to write a file that
+   * does not exist. Once close is called on the output stream, the file will be completed. Append
+   * or update of a completed file is currently not supported.
+   *
+   * @param path the Tachyon path of the file
+   * @param options the set of options specific to this operation
+   * @return an output stream to write the file
+   * @throws InvalidPathException if the provided path is invalid
+   * @throws FileAlreadyExistException if the file being written to already exists
+   * @throws BlockInfoException if the provided block size is invalid
+   * @throws IOException if the file already exists or if the stream cannot be opened
+   */
+  public FileOutStream getOutStream(TachyonURI path, ClientOptions options) throws IOException,
+      InvalidPathException, FileAlreadyExistException, BlockInfoException {
+    long fileId = super.create(path, options.getBlockSize(), true);
+    return new FileOutStream(fileId, options);
+  }
+
+  /**
+   * Alternative way to get a FileOutStream to a file that has already been created. This should
+   * not be used and will be removed in a later release.
+   *
+   * @see #getOutStream(TachyonURI path, ClientOptions options)
+   */
+  // TODO(calvin): We should remove this when the TachyonFS code is fully deprecated.
+  @Deprecated
+  public FileOutStream getOutStream(long fileId, ClientOptions options) throws IOException {
+    return new FileOutStream(fileId, options);
+  }
+
+  /**
+   * Convenience method to mkdirs recursively.
+   *
+   * @param path the Tachyon path of the folder to create recursively
+   * @return true if the directory is created
+   * @throws FileAlreadyExistException if the path already exists or a parent is a file
+   * @throws InvalidPathException if the path is invalid
+   * @throws IOException if the master cannot create the folder
+   */
+  public boolean mkdirs(TachyonURI path) throws FileAlreadyExistException, InvalidPathException,
       IOException {
+    return mkdirs(path, true);
+  }
+
+  // TODO: Move this to lineage client
+  public void reportLostFile(TachyonFile file) throws IOException, FileDoesNotExistException {
     FileSystemMasterClient masterClient = mContext.acquireMasterClient();
     try {
-      masterClient.free(file.getFileId(), recursive);
-      LOG.info("Removed file " + file.getFileId() + " from Tachyon Storage");
+      masterClient.reportLostFile(file.getFileId());
     } finally {
       mContext.releaseMasterClient(masterClient);
     }
   }
 
-  /**
-   * {@inheritDoc}
-   *
-   * The file info is a snapshot of the file metadata, and the locations, last modified time, and
-   * path are possibly inconsistent.
-   */
-  @Override
-  public FileInfo getInfo(TachyonFile file) throws IOException {
+  // TODO: Move this to lineage client
+  public void requestFilesInDependency(int depId) throws IOException,
+      DependencyDoesNotExistException {
     FileSystemMasterClient masterClient = mContext.acquireMasterClient();
     try {
-      return masterClient.getFileInfo(file.getFileId());
-    } catch (FileDoesNotExistException e) {
-      return null;
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   *
-   * The file infos are snapshots of the file metadata, and the locations, last modified time, and
-   * path are possibly inconsistent.
-   */
-  @Override
-  public List<FileInfo> listStatus(TachyonFile file) throws FileDoesNotExistException, IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      return masterClient.getFileInfoList(file.getFileId());
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
-  }
-
-
-  /**
-   * {@inheritDoc}
-   *
-   * To add the data into Tachyon space perform an operation with the cache option specified, for
-   * example the load command of the Tachyon Shell.
-   */
-  @Override
-  public long loadFileInfoFromUfs(TachyonURI path, TachyonURI ufsPath, boolean recursive)
-      throws FileDoesNotExistException, IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      long fileId = masterClient.loadFileInfoFromUfs(path.getPath(), ufsPath.toString(), recursive);
-      LOG.info(
-          "Loaded file " + path.getPath() + " from " + ufsPath + (recursive ? " recursively" : ""));
-      return fileId;
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public boolean mkdirs(TachyonURI path, boolean recursive) throws InvalidPathException,
-      IOException, FileAlreadyExistException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      // TODO: Change this RPC's arguments
-      boolean result = masterClient.createDirectory(path.getPath(), true);
-      if (result) {
-        LOG.info("Created directory " + path.getPath());
-      }
-      return result;
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public TachyonFile open(TachyonURI path) throws InvalidPathException, IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      return new TachyonFile(masterClient.getFileId(path.getPath()));
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   *
-   * The pin status is propagated asynchronously from this method call on the worker heartbeats.
-   */
-  @Override
-  public void setPin(TachyonFile file, boolean pinned) throws FileDoesNotExistException,
-      IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      masterClient.setPinned(file.getFileId(), pinned);
-      LOG.info(pinned ? "Pinned" : "Unpinned" + " file " + file.getFileId());
-    } finally {
-      mContext.releaseMasterClient(masterClient);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public boolean rename(TachyonFile src, TachyonURI dst) throws FileDoesNotExistException,
-      IOException {
-    FileSystemMasterClient masterClient = mContext.acquireMasterClient();
-    try {
-      boolean result = masterClient.renameFile(src.getFileId(), dst.getPath());
-      if (result) {
-        LOG.info("Renamed file " + src.getFileId() + " to " + dst.getPath());
-      }
-      return result;
+      masterClient.requestFilesInDependency(depId);
     } finally {
       mContext.releaseMasterClient(masterClient);
     }
