+++ examples/src/main/java/tachyon/examples/BasicCheckpoint.java
@@ -35,7 +35,6 @@ import tachyon.client.TachyonFile;
 import tachyon.client.TachyonFS;
 import tachyon.client.ClientOptions;
 import tachyon.client.file.FileInStream;
-import tachyon.client.file.StreamingTachyonFileSystem;
 import tachyon.client.file.TachyonFileSystem;
 import tachyon.conf.TachyonConf;
 import tachyon.master.file.meta.DependencyType;
@@ -111,7 +110,7 @@ public class BasicCheckpoint implements Callable<Boolean> {
       TachyonURI filePath = new TachyonURI(mFileFolder + "/part-" + i);
       LOG.debug("Writing data to {}", filePath);
       OutputStream os =
-          StreamingTachyonFileSystem.get().getOutStream(filePath, ClientOptions.defaults());
+          TachyonFileSystem.get().getOutStream(filePath, ClientOptions.defaults());
       os.write(buf.array());
       os.close();
     }
