+++ integration-tests/src/test/java/tachyon/client/BlockInStreamIntegrationTest.java
@@ -27,7 +27,7 @@ import org.junit.Test;
 
 import tachyon.Constants;
 import tachyon.client.file.FileInStream;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.client.file.TachyonFile;
 import tachyon.conf.TachyonConf;
 import tachyon.master.LocalTachyonCluster;
@@ -44,7 +44,7 @@ public class BlockInStreamIntegrationTest {
   private static final int DELTA = 33;
 
   private static LocalTachyonCluster sLocalTachyonCluster;
-  private static StreamingTachyonFileSystem sTfs;
+  private static TachyonFileSystem sTfs;
   private static TachyonConf sTachyonConf;
   private static ClientOptions sWriteBoth;
   private static ClientOptions sWriteTachyon;
