+++ integration-tests/src/test/java/tachyon/client/FileInStreamIntegrationTest.java
@@ -29,7 +29,7 @@ import org.junit.rules.ExpectedException;
 
 import tachyon.Constants;
 import tachyon.client.file.FileInStream;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.client.file.TachyonFile;
 import tachyon.conf.TachyonConf;
 import tachyon.master.LocalTachyonCluster;
@@ -47,7 +47,7 @@ public class FileInStreamIntegrationTest {
   private static final int DELTA = BLOCK_SIZE / 2;
 
   private static LocalTachyonCluster sLocalTachyonCluster = null;
-  private static StreamingTachyonFileSystem sTfs = null;
+  private static TachyonFileSystem sTfs = null;
   private static TachyonConf sTachyonConf;
   private static ClientOptions sWriteBoth;
   private static ClientOptions sWriteTachyon;
