+++ integration-tests/src/test/java/tachyon/client/FileOutStreamIntegrationTest.java
@@ -35,7 +35,7 @@ import tachyon.IntegrationTestConstants;
 import tachyon.TachyonURI;
 import tachyon.client.file.FileInStream;
 import tachyon.client.file.FileOutStream;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.client.file.TachyonFile;
 import tachyon.conf.TachyonConf;
 import tachyon.master.LocalTachyonCluster;
@@ -63,7 +63,7 @@ public class FileOutStreamIntegrationTest {
   private static ClientOptions sWriteTachyon;
   private static ClientOptions sWriteUnderStore;
 
-  private StreamingTachyonFileSystem mTfs = null;
+  private TachyonFileSystem mTfs = null;
   private TachyonConf mMasterTachyonConf;
   // If true, clients will write directly to the local file.
   private final boolean mEnableLocalWrite;
