+++ integration-tests/src/test/java/tachyon/client/TachyonFileSystemIntegrationTest.java
@@ -29,7 +29,7 @@ import org.junit.rules.ExpectedException;
 
 import tachyon.Constants;
 import tachyon.TachyonURI;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.client.file.TachyonFile;
 import tachyon.conf.TachyonConf;
 import tachyon.master.LocalTachyonCluster;
@@ -46,7 +46,7 @@ public class TachyonFileSystemIntegrationTest {
   private static LocalTachyonCluster sLocalTachyonCluster = null;
   private static String sHost = null;
   private static int sPort = -1;
-  private static StreamingTachyonFileSystem sTfs = null;
+  private static TachyonFileSystem sTfs = null;
   private static ClientOptions sReadCache;
   private static ClientOptions sWriteBoth;
   private TachyonConf mMasterTachyonConf;
