+++ integration-tests/src/test/java/tachyon/client/table/RawTableIntegrationTest.java
@@ -19,9 +19,7 @@ import org.junit.After;
 import org.junit.Before;
 
 import tachyon.Constants;
-import tachyon.client.file.FileInStream;
-import tachyon.client.file.FileOutStream;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.master.LocalTachyonCluster;
 
 /**
@@ -29,7 +27,7 @@ import tachyon.master.LocalTachyonCluster;
  */
 public class RawTableIntegrationTest {
   private LocalTachyonCluster mLocalTachyonCluster = null;
-  private StreamingTachyonFileSystem mTfs = null;
+  private TachyonFileSystem mTfs = null;
   private int mMaxCols = 1000;
 
   @After
