+++ integration-tests/src/test/java/tachyon/hadoop/TFSStatisticsTest.java
@@ -31,7 +31,7 @@ import org.junit.Test;
 import tachyon.client.TachyonStorageType;
 import tachyon.client.TachyonFSTestUtils;
 import tachyon.client.UnderStorageType;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.master.LocalTachyonCluster;
 
 /**
@@ -54,7 +54,7 @@ public class TFSStatisticsTest {
     sLocalTachyonCluster = new LocalTachyonCluster(10000, 1000, BLOCK_SIZE);
     sLocalTachyonCluster.start();
 
-    StreamingTachyonFileSystem tachyonFS = sLocalTachyonCluster.getClient();
+    TachyonFileSystem tachyonFS = sLocalTachyonCluster.getClient();
     TachyonFSTestUtils.createByteFile(tachyonFS, "/testFile-read", TachyonStorageType.STORE,
         UnderStorageType.PERSIST, FILE_LEN);
