+++ integration-tests/src/test/java/tachyon/master/JournalShutdownIntegrationTest.java
@@ -29,7 +29,7 @@ import org.junit.Test;
 import tachyon.Constants;
 import tachyon.TachyonURI;
 import tachyon.client.ClientOptions;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.conf.TachyonConf;
 import tachyon.master.file.FileSystemMaster;
 import tachyon.thrift.FileDoesNotExistException;
@@ -51,9 +51,9 @@ public class JournalShutdownIntegrationTest {
     private int mSuccessNum = 0;
 
     private final int mOpType; // 0:create file; 1:create raw table.
-    private final StreamingTachyonFileSystem mTfs;
+    private final TachyonFileSystem mTfs;
 
-    public ClientThread(int opType, StreamingTachyonFileSystem tfs) {
+    public ClientThread(int opType, TachyonFileSystem tfs) {
       mOpType = opType;
       mTfs = tfs;
     }
