+++ shell/src/test/java/tachyon/shell/TFsShellUtilsTest.java
@@ -34,7 +34,7 @@ import tachyon.TachyonURI;
 import tachyon.client.TachyonStorageType;
 import tachyon.client.TachyonFSTestUtils;
 import tachyon.client.UnderStorageType;
-import tachyon.client.file.StreamingTachyonFileSystem;
+import tachyon.client.file.TachyonFileSystem;
 import tachyon.client.file.TachyonFile;
 import tachyon.conf.TachyonConf;
 import tachyon.master.LocalTachyonCluster;
@@ -49,7 +49,7 @@ import tachyon.thrift.InvalidPathException;
 public class TFsShellUtilsTest {
   private static final int SIZE_BYTES = Constants.MB * 10;
   private LocalTachyonCluster mLocalTachyonCluster = null;
-  private StreamingTachyonFileSystem mTfs = null;
+  private TachyonFileSystem mTfs = null;
 
   @After
   public final void after() throws Exception {
@@ -101,7 +101,7 @@ public class TFsShellUtilsTest {
     return resetTachyonFileHierarchy(mTfs);
   }
 
-  public static String resetTachyonFileHierarchy(StreamingTachyonFileSystem tfs) throws IOException,
+  public static String resetTachyonFileHierarchy(TachyonFileSystem tfs) throws IOException,
       TException {
     /**
      * Generate such local structure /testWildCards
